chore: rearrange the deck chairs
