+++ .github/workflows/prepare.yaml
@@ -13,7 +13,7 @@ jobs:
       - uses: actions/setup-go@v5
         with:
             go-version: '1.22'
-      - name: Release Verseion
+      - name: Release Version
         run: echo RELEASE_VERSION=$(echo $GITHUB_REF_NAME | grep -Eo '[0-9]+\.[0-9]+\.[0-9]+.*$') >> $GITHUB_ENV
       - name: Version Check
         run: |
